New script EnemyModel
Created new script enemy model, that contains only model related functions. Enemy script now contains reference to the enemy model, is smaller and easier to read.
