+++ New Unity Project/Assets/ScriptableObjects/EnemiesSO.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ac0505ed77b7c24fb460368b69ef4f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
