+++ New Unity Project/Assets/ScriptableObjects/EnemiesSO/BasicEnemy.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d13ab1b7943ac5647bbabf1e8b3cab7f, type: 3}
+  m_Name: BasicEnemy
+  m_EditorClassIdentifier: 
+  enemyName: Basic
+  health: 50
+  enemySize: 1
