+++ New Unity Project/Assets/ScriptableObjects/EnemiesSO/BasicEnemy.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d65f80cd45fb06409e80921e6392b03
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
