+++ New Unity Project/Assets/Scripts/Enemy.cs
@@ -6,28 +6,16 @@ using UnityEngine;
 public class Enemy : MonoBehaviour
 {
     [SerializeField] public EnemySO enemySO;
-    private Color hitColor = Color.red;
-    private Color baseColor = Color.white;
-
-    float health;
-
-    Rigidbody rigidbody;
-    Renderer renderer;
-    Animator animator;
-
-    private float hitColorChangeTimer = 0.5f;
-    private float enemyDeathFallTimer = 1.5f;
+    [SerializeField] public EnemyModel enemyModel;
+    
+    public float health;
 
     public delegate void OnHealthChange();
     public static event OnHealthChange HealthDrop;
 
-
     private void Start()
     {
-        rigidbody = GetComponent<Rigidbody>();
-        rigidbody.useGravity = false;
-        renderer = GetComponent<Renderer>();
-        animator = GetComponent<Animator>();
+        enemyModel = new EnemyModel();
         health = enemySO.health;
     }
     public void TakeDamage(float amount)
@@ -37,30 +25,18 @@ public class Enemy : MonoBehaviour
        
         if (health <= 0f)
         {
-            animator.enabled = false;
-            EnemyDeath();
+            enemyModel.EnemyDeath();
             Debug.Log("Enemy died!!");
-            
-
         }
         if(health > 0f)
         {
-            StartCoroutine(ChangeTargetColorOnHit());
+            StartCoroutine(enemyModel.ChangeTargetColorOnHit());
         }
     }
 
-    private void EnemyDeath()
-    {
-        rigidbody.useGravity = true;
-    }
+    
 
-    public IEnumerator ChangeTargetColorOnHit()
-    {
-        renderer.material.color = hitColor;
-        yield return new WaitForSeconds(hitColorChangeTimer);
-        renderer.material.color = baseColor;
-    }
-     
+   
     
     
 }
