+++ New Unity Project/Assets/Scripts/EnemyModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2b4948c77e342543afcbcfaf0effd70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
